+++ Toontown Empire/src/dependencies/panda/include/pnmPainter.h
@@ -16,7 +16,6 @@
 #define PNMPAINTER_H
 
 #include "pandabase.h"
-#include "pnmBrush.h"
 
 class PNMImage;
 
@@ -41,17 +40,17 @@ PUBLISHED:
   INLINE void set_fill(PNMBrush *fill);
   INLINE PNMBrush *get_fill() const;
 
-  INLINE void draw_point(float x, float y);
-  void draw_line(float xa, float ya, float xb, float yb);
-  void draw_rectangle(float xa, float ya, float xb, float yb);
+  INLINE void draw_point(double x, double y);
+  void draw_line(double xa, double ya, double xb, double yb);
+  void draw_rectangle(double xa, double ya, double xb, double yb);
 
 private:
-  INLINE void draw_hline_point(int x, float xa, float ya,
-                               float xd, float yd,
-                               float pixel_scale);
-  INLINE void draw_vline_point(int y, float xa, float ya,
-                               float xd, float yd,
-                               float pixel_scale);
+  INLINE void draw_hline_point(int x, double xa, double ya, 
+                               double xd, double yd,
+                               double pixel_scale);
+  INLINE void draw_vline_point(int y, double xa, double ya, 
+                               double xd, double yd,
+                               double pixel_scale);
 
 private:
   PNMImage &_image;
